feat: refactor Header and Sidebar into modular layout components (#465)
- Extract AppShell inline header JSX into frontend/src/components/layout/Header.tsx
- Move Sidebar implementation to frontend/src/components/layout/Sidebar.tsx
- Keep components/Sidebar.tsx as a backward-compat re-export shim
- Header is now a focused global-chrome component: logo, StepperNav (learning mode), notifications bell, user name, logout
- All page-level navigation links (班級管理, 作業, 生字, 記錄, 進度...) removed from header — they live in Sidebar
- App.tsx AppShell reduced from ~280 lines of inline JSX to a clean composition
- Role-based nav items remain data-driven arrays in Sidebar (no change to that logic)
- All existing functionality preserved; build passes

via [Happy](https://happy.engineering)
